Bug fixes.
Fixed a bug where CollectorUpdate was returning instead of continuing when a flag set condition is met during the update loop.
Fixed a bug where dtlim was being set incorrectly.
Fixed a bug where the pause/resume time elapsed was discarded from the for-conditional expression.
Fixed a bug where after the first pass items would never be processed again due to not looping the index around.
Fixed a bug where only the first item would be processed on a single pass.
Fixed a bug where the pending flag was being set instead of cleared.
Fixed a bug where dt was being improperly calculated due to not taking a time reference properly.